Create Driver area, fix check used vehicle at create

+++ TransportManagement/Models/Vehicle/VehicleViewModel.cs
@@ -12,14 +12,14 @@ namespace TransportManagement.Models.Vehicle
         private string _vehicleName;
         private string _brandId;
         private string _brandName;
-        private int _vehiclePayload;
+        private decimal _vehiclePayload;
         private bool _isInUse;
         private bool _isAvailable;
 
         public int VehicleId { get => _vehicleId; set => _vehicleId = value; }
         public string LicensePlate { get => _licensePlate; set => _licensePlate = value; }
         public string VehicleName { get => _vehicleName; set => _vehicleName = value; }
-        public int VehiclePayload { get => _vehiclePayload; set => _vehiclePayload = value; }
+        public decimal VehiclePayload { get => _vehiclePayload; set => _vehiclePayload = value; }
         public bool IsInUse { get => _isInUse; set => _isInUse = value; }
         public bool IsAvailable { get => _isAvailable; set => _isAvailable = value; }
         public string BrandName { get => _brandName; set => _brandName = value; }
